backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
chromadb==0.5.3
google-generativeai==0.7.2
groq==0.9.0
python-dotenv==1.0.1
tqdm==4.66.4
